Initial implementation of Linux support

+++ polymer/types.h
@@ -22,7 +22,7 @@ using u16 = uint16_t;
 using u32 = uint32_t;
 using u64 = uint64_t;
 
-using wchar = u16;
+using wchar = u32;
 
 struct String {
   char* data;
